+++ tests/unit/drivers/prompt/test_ollama_prompt_driver.py
@@ -16,19 +16,6 @@ class TestOllamaPromptDriver:
         "required": ["foo"],
         "type": "object",
     }
-    OLLAMA_STRUCTURED_OUTPUT_TOOL = {
-        "function": {
-            "description": "Used to provide the final response which ends this conversation.",
-            "name": "StructuredOutputTool_provide_output",
-            "parameters": {
-                "additionalProperties": False,
-                "properties": {"foo": {"type": "string"}},
-                "required": ["foo"],
-                "type": "object",
-            },
-        },
-        "type": "function",
-    }
     OLLAMA_TOOLS = [
         {
             "function": {
@@ -232,22 +219,19 @@ class TestOllamaPromptDriver:
         assert OllamaPromptDriver(model="llama")
 
     @pytest.mark.parametrize("use_native_tools", [True, False])
-    @pytest.mark.parametrize("use_structured_output", [True, False])
-    @pytest.mark.parametrize("structured_output_strategy", ["native", "tool", "foo"])
+    @pytest.mark.parametrize("structured_output_strategy", ["native", "tool", "rule", "foo"])
     def test_try_run(
         self,
         mock_client,
         prompt_stack,
         messages,
         use_native_tools,
-        use_structured_output,
         structured_output_strategy,
     ):
         # Given
         driver = OllamaPromptDriver(
             model="llama",
             use_native_tools=use_native_tools,
-            use_structured_output=use_structured_output,
             structured_output_strategy=structured_output_strategy,
             extra_params={"foo": "bar"},
         )
@@ -265,20 +249,11 @@ class TestOllamaPromptDriver:
                 "num_predict": driver.max_tokens,
             },
             **{
-                "tools": [
-                    *self.OLLAMA_TOOLS,
-                    *(
-                        [self.OLLAMA_STRUCTURED_OUTPUT_TOOL]
-                        if use_structured_output and structured_output_strategy == "tool"
-                        else []
-                    ),
-                ]
+                "tools": self.OLLAMA_TOOLS,
             }
             if use_native_tools
             else {},
-            **{"format": self.OLLAMA_STRUCTURED_OUTPUT_SCHEMA}
-            if use_structured_output and structured_output_strategy == "native"
-            else {},
+            **{"format": self.OLLAMA_STRUCTURED_OUTPUT_SCHEMA} if structured_output_strategy == "native" else {},
             foo="bar",
         )
         assert isinstance(message.value[0], TextArtifact)
@@ -290,15 +265,13 @@ class TestOllamaPromptDriver:
         assert message.value[1].value.input == {"foo": "bar"}
 
     @pytest.mark.parametrize("use_native_tools", [True, False])
-    @pytest.mark.parametrize("use_structured_output", [True, False])
-    @pytest.mark.parametrize("structured_output_strategy", ["native", "tool", "foo"])
+    @pytest.mark.parametrize("structured_output_strategy", ["native", "tool", "rule", "foo"])
     def test_try_stream_run(
         self,
         mock_stream_client,
         prompt_stack,
         messages,
         use_native_tools,
-        use_structured_output,
         structured_output_strategy,
     ):
         # Given
@@ -306,7 +279,6 @@ class TestOllamaPromptDriver:
             model="llama",
             stream=True,
             use_native_tools=use_native_tools,
-            use_structured_output=use_structured_output,
             structured_output_strategy=structured_output_strategy,
             extra_params={"foo": "bar"},
         )
@@ -319,9 +291,7 @@ class TestOllamaPromptDriver:
             messages=messages,
             model=driver.model,
             options={"temperature": driver.temperature, "stop": [], "num_predict": driver.max_tokens},
-            **{"format": self.OLLAMA_STRUCTURED_OUTPUT_SCHEMA}
-            if use_structured_output and structured_output_strategy == "native"
-            else {},
+            **{"format": self.OLLAMA_STRUCTURED_OUTPUT_SCHEMA} if structured_output_strategy == "native" else {},
             stream=True,
             foo="bar",
         )